haku-app/src/main/java/fi/vm/sade/haku/oppija/ui/controller/RootController.java
/*
 * Copyright (c) 2012 The Finnish Board of Education - Opetushallitus
 *
 * This program is free software:  Licensed under the EUPL, Version 1.1 or - as
 * soon as they will be approved by the European Commission - subsequent versions
 * of the EUPL (the "Licence");
 *
 * You may not use this work except in compliance with the Licence.
 * You may obtain a copy of the Licence at: http://www.osor.eu/eupl/
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
 * European Union Public Licence for more details.
 */

package fi.vm.sade.haku.oppija.ui.controller;

import org.glassfish.jersey.server.mvc.Viewable;
import org.springframework.context.annotation.Profile;
import org.springframework.stereotype.Controller;

import javax.ws.rs.GET;
import javax.ws.rs.Path;
import javax.ws.rs.Produces;
import javax.ws.rs.core.MediaType;

@Path("/")
@Controller
@Profile(value = {"dev", "it", "devluokka"})
public class RootController {

    private static final String CHARSET_UTF_8 = ";charset=UTF-8";
    public static final String INDEX_VIEW = "/index";

    @GET
    @Produces(MediaType.TEXT_HTML + CHARSET_UTF_8)
    public Viewable getFrontPage() {
        return new Viewable(INDEX_VIEW);
    }
}
